+++ BusinessLogic/Commands/Element/IDeleteStoryElementInSpace.cs
@@ -1,0 +1,5 @@
+namespace BusinessLogic.Commands.Element;
+
+public interface IDeleteStoryElementInSpace : IUndoCommand
+{
+}