+++ Generator/ATF/ILearningSpaceJson.cs
@@ -23,4 +23,5 @@ public interface ILearningSpaceJson
     string SpaceTemplateStyle { get; set; }
     string SpaceName { get; set; }
     int RequiredPointsToComplete { get; set; }
+    ISpaceStoryJson SpaceStory { get; set; }
 }