+++ Presentation/PresentationLogic/LearningContent/Story/IStoryContentViewModel.cs
@@ -1,0 +1,6 @@
+namespace Presentation.PresentationLogic.LearningContent.Story;
+
+public interface IStoryContentViewModel : ILearningContentViewModel
+{
+    public List<string> StoryText { get; set; }
+}